+++ rfaas/lib/resources.cpp
@@ -15,7 +15,7 @@ namespace rfaas {
     memory(-1)
   {}
 
-  server_data::server_data(const std::string & node, const std::string & ip, int32_t port, int16_t cores, int32_t memory):
+  server_data::server_data(const std::string & node, const std::string & ip, int32_t port, int16_t cores, int64_t memory):
     port(port),
     cores(cores),
     memory(memory),
